fix: preserve OpenRouter metric preference
