+++ Tests/CodexBarTests/StatusItemAnimationTests.swift
@@ -888,96 +888,6 @@ struct StatusItemAnimationTests {
     }
 
     @Test
-    func `menu bar display text uses open router balance`() {
-        let settings = SettingsStore(
-            configStore: testConfigStore(suiteName: "StatusItemAnimationTests-openrouter-balance"),
-            zaiTokenStore: NoopZaiTokenStore())
-        settings.statusChecksEnabled = false
-        settings.refreshFrequency = .manual
-        settings.mergeIcons = true
-        settings.selectedMenuProvider = .openrouter
-        settings.menuBarDisplayMode = .both
-        settings.usageBarsShowUsed = true
-
-        let registry = ProviderRegistry.shared
-        if let openRouterMeta = registry.metadata[.openrouter] {
-            settings.setProviderEnabled(provider: .openrouter, metadata: openRouterMeta, enabled: true)
-        }
-
-        let fetcher = UsageFetcher()
-        let store = UsageStore(fetcher: fetcher, browserDetection: BrowserDetection(cacheTTL: 0), settings: settings)
-        let controller = StatusItemController(
-            store: store,
-            settings: settings,
-            account: fetcher.loadAccountInfo(),
-            updater: DisabledUpdaterController(),
-            preferencesSelection: PreferencesSelection(),
-            statusBar: self.makeStatusBarForTesting())
-
-        let snapshot = OpenRouterUsageSnapshot(
-            totalCredits: 50,
-            totalUsage: 37.66,
-            balance: 12.34,
-            usedPercent: 75.32,
-            keyLimit: 20,
-            keyUsage: 5,
-            rateLimit: nil,
-            updatedAt: Date()).toUsageSnapshot()
-
-        store._setSnapshotForTesting(snapshot, provider: .openrouter)
-        store._setErrorForTesting(nil, provider: .openrouter)
-
-        let displayText = controller.menuBarDisplayText(for: .openrouter, snapshot: snapshot)
-
-        #expect(displayText == "$12.34")
-    }
-
-    @Test
-    func `menu bar display text uses deepseek balance`() {
-        let settings = SettingsStore(
-            configStore: testConfigStore(suiteName: "StatusItemAnimationTests-deepseek-balance"),
-            zaiTokenStore: NoopZaiTokenStore(),
-            syntheticTokenStore: NoopSyntheticTokenStore())
-        settings.statusChecksEnabled = false
-        settings.refreshFrequency = .manual
-        settings.mergeIcons = true
-        settings.selectedMenuProvider = .deepseek
-        settings.menuBarDisplayMode = .both
-        settings.usageBarsShowUsed = true
-
-        let registry = ProviderRegistry.shared
-        if let deepSeekMeta = registry.metadata[.deepseek] {
-            settings.setProviderEnabled(provider: .deepseek, metadata: deepSeekMeta, enabled: true)
-        }
-
-        let fetcher = UsageFetcher()
-        let store = UsageStore(fetcher: fetcher, browserDetection: BrowserDetection(cacheTTL: 0), settings: settings)
-        let controller = StatusItemController(
-            store: store,
-            settings: settings,
-            account: fetcher.loadAccountInfo(),
-            updater: DisabledUpdaterController(),
-            preferencesSelection: PreferencesSelection(),
-            statusBar: self.makeStatusBarForTesting())
-
-        let snapshot = UsageSnapshot(
-            primary: RateWindow(
-                usedPercent: 0,
-                windowMinutes: nil,
-                resetsAt: nil,
-                resetDescription: "$9.32 (Paid: $9.32 / Granted: $0.00)"),
-            secondary: nil,
-            updatedAt: Date())
-
-        store._setSnapshotForTesting(snapshot, provider: .deepseek)
-        store._setErrorForTesting(nil, provider: .deepseek)
-
-        let displayText = controller.menuBarDisplayText(for: .deepseek, snapshot: snapshot)
-
-        #expect(displayText == "$9.32")
-    }
-
-    @Test
     func `menu bar display text shows zero percent for kilo zero total edge`() {
         let settings = SettingsStore(
             configStore: testConfigStore(suiteName: "StatusItemAnimationTests-kilo-zero-edge"),
